new version

+++ frontend/modules/checklist/views/default/checklist_items.php
@@ -9,14 +9,18 @@ use yii\widgets\ActiveForm;
 ?>
 
 <div class="checklist-form" data-target="<?= $checklist_id ?>">
-    <? foreach ($checklist_options as $checklist_option): ?>
-        <label>
-            <input type="checkbox" value="1"/>
-            <span><?= $checklist_option['name'] ?></span>
-            <a href="#!" class="delete_item" data-target="<?= $checklist_option['id'] ?>">
-                <i class="material-icons">clear</i>
-            </a>
-        </label>
-    <? endforeach; ?>
+    <? if ($checklist_options):?>
+        <? foreach ($checklist_options as $checklist_option): ?>
+            <label>
+                <input type="checkbox" value="1"/>
+                <span><?= $checklist_option['name'] ?></span>
+                <a href="#!" class="delete_item" data-target="<?= $checklist_option['id'] ?>">
+                    <i class="material-icons">clear</i>
+                </a>
+            </label>
+        <? endforeach; ?>
+    <? else: ?>
+        <div class="empty-checklist">Пустой чек-лист</div>
+    <? endif; ?>
 </div>
